[FEATURE] Allow preview URLs for single pages
When creating a preview URL for a single page, access will only
be granted for this specific page. To choose a page the backend
module makes use of the page tree.

It would be possible to integrate a button in the page module for this
as well. This will be implementen in a later patch.